DDF-1172 Improve code coverage and correct issues
Improved code coverage for fail states and exception blocks, and also
added test cases which cover the code in the checkForMetaTypesForService
method. In order to do this, I had to put in a package-private setter
for serviceTracker. Also changed over most test strings to be defined as
constants, and reduced the use of assertNotNull(..), instead using the
more verbose assertThat(..).