DDF-1472: Updated Error Handling to return 401
Since this version can also process saml
assertions in headers it is necessary to return a 401 error
when an invalid assertion is found instead of invalidating
non-existant cookies and redirecting them to try again. In
the unlikely event that the saml assertion is truely broken,
this would cause a processing loop while the client continued
to try and resend the header and while getting back updates to
expire non-existant cookies.

 - Updated WebSSOFilter to return a 401
 - Updates SAMLAssertionHandler to return 401
 - Removed leftover comment lines
 - encodeSAML and decodeSAML methods were migrated to RestSecurity,
   so the encode and decode methods in this class were removed and
   changed to use the methods defined in RestSecurity.